+++ packages/google-cloud-geminidataanalytics/src/v1alpha/data_chat_service_client.ts
@@ -305,6 +305,7 @@ export class DataChatServiceClient {
       'chat',
       'createConversation',
       'deleteConversation',
+      'updateConversation',
       'getConversation',
       'listConversations',
       'listMessages',
@@ -741,6 +742,162 @@ export class DataChatServiceClient {
       });
   }
   /**
+   * Updates a conversation.
+   *
+   * @param {Object} request
+   *   The request object that will be sent.
+   * @param {google.cloud.geminidataanalytics.v1alpha.Conversation} request.conversation
+   *   Required. The resource being updated.
+   * @param {google.protobuf.FieldMask} [request.updateMask]
+   *   Optional. Field mask is used to specify the fields to be overwritten in the
+   *   Conversation resource by the update.
+   *   The fields specified in the update_mask are relative to the resource, not
+   *   the full request. A field will be overwritten if it is in the mask. If the
+   *   user does not provide a mask then all fields with non-default values
+   *   present in the request will be overwritten. If a wildcard mask is provided,
+   *   all fields will be overwritten.
+   * @param {string} [request.requestId]
+   *   Optional. An optional request ID to identify requests. Specify a unique
+   *   request ID so that if you must retry your request, the server will know to
+   *   ignore the request if it has already been completed. The server will
+   *   guarantee that for at least 60 minutes since the first request.
+   * @param {object} [options]
+   *   Call options. See {@link https://googleapis.dev/nodejs/google-gax/latest/interfaces/CallOptions.html|CallOptions} for more details.
+   * @returns {Promise} - The promise which resolves to an array.
+   *   The first element of the array is an object representing {@link protos.google.cloud.geminidataanalytics.v1alpha.Conversation|Conversation}.
+   *   Please see the {@link https://github.com/googleapis/gax-nodejs/blob/master/client-libraries.md#regular-methods | documentation }
+   *   for more details and examples.
+   * @example <caption>include:samples/generated/v1alpha/data_chat_service.update_conversation.js</caption>
+   * region_tag:geminidataanalytics_v1alpha_generated_DataChatService_UpdateConversation_async
+   */
+  updateConversation(
+    request?: protos.google.cloud.geminidataanalytics.v1alpha.IUpdateConversationRequest,
+    options?: CallOptions,
+  ): Promise<
+    [
+      protos.google.cloud.geminidataanalytics.v1alpha.IConversation,
+      (
+        | protos.google.cloud.geminidataanalytics.v1alpha.IUpdateConversationRequest
+        | undefined
+      ),
+      {} | undefined,
+    ]
+  >;
+  updateConversation(
+    request: protos.google.cloud.geminidataanalytics.v1alpha.IUpdateConversationRequest,
+    options: CallOptions,
+    callback: Callback<
+      protos.google.cloud.geminidataanalytics.v1alpha.IConversation,
+      | protos.google.cloud.geminidataanalytics.v1alpha.IUpdateConversationRequest
+      | null
+      | undefined,
+      {} | null | undefined
+    >,
+  ): void;
+  updateConversation(
+    request: protos.google.cloud.geminidataanalytics.v1alpha.IUpdateConversationRequest,
+    callback: Callback<
+      protos.google.cloud.geminidataanalytics.v1alpha.IConversation,
+      | protos.google.cloud.geminidataanalytics.v1alpha.IUpdateConversationRequest
+      | null
+      | undefined,
+      {} | null | undefined
+    >,
+  ): void;
+  updateConversation(
+    request?: protos.google.cloud.geminidataanalytics.v1alpha.IUpdateConversationRequest,
+    optionsOrCallback?:
+      | CallOptions
+      | Callback<
+          protos.google.cloud.geminidataanalytics.v1alpha.IConversation,
+          | protos.google.cloud.geminidataanalytics.v1alpha.IUpdateConversationRequest
+          | null
+          | undefined,
+          {} | null | undefined
+        >,
+    callback?: Callback<
+      protos.google.cloud.geminidataanalytics.v1alpha.IConversation,
+      | protos.google.cloud.geminidataanalytics.v1alpha.IUpdateConversationRequest
+      | null
+      | undefined,
+      {} | null | undefined
+    >,
+  ): Promise<
+    [
+      protos.google.cloud.geminidataanalytics.v1alpha.IConversation,
+      (
+        | protos.google.cloud.geminidataanalytics.v1alpha.IUpdateConversationRequest
+        | undefined
+      ),
+      {} | undefined,
+    ]
+  > | void {
+    request = request || {};
+    let options: CallOptions;
+    if (typeof optionsOrCallback === 'function' && callback === undefined) {
+      callback = optionsOrCallback;
+      options = {};
+    } else {
+      options = optionsOrCallback as CallOptions;
+    }
+    options = options || {};
+    options.otherArgs = options.otherArgs || {};
+    options.otherArgs.headers = options.otherArgs.headers || {};
+    options.otherArgs.headers['x-goog-request-params'] =
+      this._gaxModule.routingHeader.fromParams({
+        'conversation.name': request.conversation!.name ?? '',
+      });
+    this.initialize().catch(err => {
+      throw err;
+    });
+    this._log.info('updateConversation request %j', request);
+    const wrappedCallback:
+      | Callback<
+          protos.google.cloud.geminidataanalytics.v1alpha.IConversation,
+          | protos.google.cloud.geminidataanalytics.v1alpha.IUpdateConversationRequest
+          | null
+          | undefined,
+          {} | null | undefined
+        >
+      | undefined = callback
+      ? (error, response, options, rawResponse) => {
+          this._log.info('updateConversation response %j', response);
+          callback!(error, response, options, rawResponse); // We verified callback above.
+        }
+      : undefined;
+    return this.innerApiCalls
+      .updateConversation(request, options, wrappedCallback)
+      ?.then(
+        ([response, options, rawResponse]: [
+          protos.google.cloud.geminidataanalytics.v1alpha.IConversation,
+          (
+            | protos.google.cloud.geminidataanalytics.v1alpha.IUpdateConversationRequest
+            | undefined
+          ),
+          {} | undefined,
+        ]) => {
+          this._log.info('updateConversation response %j', response);
+          return [response, options, rawResponse];
+        },
+      )
+      .catch((error: any) => {
+        if (
+          error &&
+          'statusDetails' in error &&
+          error.statusDetails instanceof Array
+        ) {
+          const protos = this._gaxModule.protobuf.Root.fromJSON(
+            jsonProtos,
+          ) as unknown as gax.protobuf.Type;
+          error.statusDetails = decodeAnyProtosInArray(
+            error.statusDetails,
+            protos,
+          );
+        }
+        throw error;
+      });
+  }
+  /**
    * Gets details of a single conversation by using conversation id and parent.
    *
    * @param {Object} request
@@ -1040,7 +1197,7 @@ export class DataChatServiceClient {
 
   /**
    * Answers a data question by generating a stream of
-   * {@link protos.google.cloud.geminidataanalytics.v1alpha.Message|Message} objects.
+   * {@link protos.google.cloud.geminidataanalytics.v1.Message|Message} objects.
    *
    * @param {Object} request
    *   The request object that will be sent.
@@ -1059,16 +1216,31 @@ export class DataChatServiceClient {
    *   Some clients may not use GDA managed resources including
    *   conversations and agents, instead they create and manage their own
    *   conversations and agents resources.
+   * @param {google.cloud.geminidataanalytics.v1alpha.LookerSettings} [request.lookerSettings]
+   *   Optional. Looker specific settings.
    * @param {string} [request.project]
-   *   Optional. The Google Cloud project to be used for quota and billing.
+   *   Optional. Deprecated: Use `parent` field instead.
+   *   The Google Cloud project to be used for quota and billing.
    * @param {string} request.parent
    *   Required. The parent value for chat request.
    *   Pattern: `projects/{project}/locations/{location}`
    * @param {number[]} request.messages
    *   Required. Content of current conversation.
+   * @param {google.cloud.geminidataanalytics.v1alpha.Credentials} [request.credentials]
+   *   Optional. The credentials to use when calling the data source(s) specified
+   *   in the context.
+   *
+   *   This field can be used to provide credentials for various data sources.
+   *   For example, when connecting to Looker, it currently supports both OAuth
+   *   token and API key-based credentials, as described in
+   *   [Authentication with an
+   *   SDK](https://cloud.google.com/looker/docs/api-auth#authentication_with_an_sdk).
    * @param {google.cloud.geminidataanalytics.v1alpha.ChatRequest.ThinkingMode} [request.thinkingMode]
    *   Optional. The thinking mode to use for the agent loop.
    *   Defaults to THINKING_MODE_UNSPECIFIED if not specified.
+   * @param {google.cloud.geminidataanalytics.v1alpha.ChatRequest.Model} [request.model]
+   *   Optional. The model to use for the agent loop when processing the request.
+   *   This setting only has an effect when context.options.model is not set.
    * @param {object} [options]
    *   Call options. See {@link https://googleapis.dev/nodejs/google-gax/latest/interfaces/CallOptions.html|CallOptions} for more details.
    * @returns {Stream}
@@ -1107,8 +1279,9 @@ export class DataChatServiceClient {
    *   Format: `projects/{project}/locations/{location}`
    * @param {number} [request.pageSize]
    *   Optional. Requested page size. Server may return fewer items than
-   *   requested. The max page size is 100. All larger page sizes will be coerced
-   *   to 100. If unspecified, server will pick 50 as an approperiate default.
+   *   requested. The max page size is `100`. All larger page sizes will be
+   *   coerced to `100`. If unspecified, server will pick `50` as an appropriate
+   *   default.
    * @param {string} [request.pageToken]
    *   Optional. A token identifying a page of results the server should return.
    * @param {string} [request.filter]
@@ -1240,8 +1413,9 @@ export class DataChatServiceClient {
    *   Format: `projects/{project}/locations/{location}`
    * @param {number} [request.pageSize]
    *   Optional. Requested page size. Server may return fewer items than
-   *   requested. The max page size is 100. All larger page sizes will be coerced
-   *   to 100. If unspecified, server will pick 50 as an approperiate default.
+   *   requested. The max page size is `100`. All larger page sizes will be
+   *   coerced to `100`. If unspecified, server will pick `50` as an appropriate
+   *   default.
    * @param {string} [request.pageToken]
    *   Optional. A token identifying a page of results the server should return.
    * @param {string} [request.filter]
@@ -1296,8 +1470,9 @@ export class DataChatServiceClient {
    *   Format: `projects/{project}/locations/{location}`
    * @param {number} [request.pageSize]
    *   Optional. Requested page size. Server may return fewer items than
-   *   requested. The max page size is 100. All larger page sizes will be coerced
-   *   to 100. If unspecified, server will pick 50 as an approperiate default.
+   *   requested. The max page size is `100`. All larger page sizes will be
+   *   coerced to `100`. If unspecified, server will pick `50` as an appropriate
+   *   default.
    * @param {string} [request.pageToken]
    *   Optional. A token identifying a page of results the server should return.
    * @param {string} [request.filter]